+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - `encoding` and `standalone` to AttributeId.
+- `new_translate`, `new_scale`, `new_rotate`, `new_rotate_at`, `new_skew_x`, `new_skew_y`
+  and `rotate_at` methods to the `Transform`.
 
 ### Changed
 - `StreamExt::parse_iri` and `StreamExt::parse_func_iri` will parse